rename package from v1auth to v1_auth
