+++ backend/internal/v1/v1_auth/routes.go
@@ -1,1 +1,1 @@
-package v1auth
+package v1_auth
