+++ backend/internal/v1/v1_auth/types.go
@@ -1,1 +1,1 @@
-package v1auth
+package v1_auth
